+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/player/InventoryCleaner.kt
@@ -434,7 +434,7 @@ object InventoryCleaner : Module("InventoryCleaner", Category.PLAYER) {
             // Have to set this manually, because it would delay all clicks until a first scheduled click was sent
             isFirstInventoryClick = false
 
-            delay(startDelay.toLong())
+            delay(startDelay.random().toLong())
         }
 
         clickNextTick(slot, button, mode, allowDuplicates)